refactor: move pure io/validation helpers inward to core, shared re-exports
